Newist intake and shoot comands

+++ src/main/java/frc/robot/subsystems/CassetteIntake.java
@@ -19,6 +19,7 @@ public class CassetteIntake extends SubsystemBase {
 
   /* Intake speed in rps (Doesn't account for mechanism gearing on it's own) */
   private static double INTAKE_SPEED = 1; // TODO: tune
+  
                                                                                                                   
   /** Creates a new CassetteIntake. */
   public CassetteIntake() {
@@ -36,7 +37,7 @@ public class CassetteIntake extends SubsystemBase {
    * Start or stop intake at constant (tested) speed
    * @param state Use -1 to invert, 0 to stop, and 1 to run forward
   */
-  public void intakeRun(int state) {
+  public void intakeRun(double state) {
     m_IntakeMotor1.set(INTAKE_SPEED * state); // Todo, this uses percent output, not velocity control as it should
     m_IntakeMotor2.set(INTAKE_SPEED * state);
   }
